agentcore-runtime/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
opensearch-py
boto3
requests-aws4auth
click
websockets
